Break forcecrc32.c into 2 modules
This is rather handier since this way you don't need to modify the C
source file to build it into another project.  This also avoids the
necessity of having to define an extra macro to disable the main()
function.
